UJS-270: Fixed delete in products, productstravelling. Fixed store name in products.

See merge request rubber-duck/uvel!534

+++ app/Http/Controllers/ProductTravellingController.php
@@ -79,6 +79,7 @@ class ProductTravellingController extends Controller
                 $travel->save();
 
                 $product = Product::find($product);
+                $product->store_id = 1;
                 $product->status = 'travelling';
                 $product->save();
 
@@ -182,13 +183,14 @@ class ProductTravellingController extends Controller
     {
         if($product){
             $originProduct = Product::find($product->product_id);
-            if(Order::where('product_id', $product->product_id)){
+            if(($order = Order::where('product_id', $product->product_id)) && $order->first()){
                 $order = Order::where('product_id', $product->product_id)->first();
                 $order->product_id = null;
                 $order->status = 'accepted';
                 $order->save();
                 $originProduct->delete();
             } else {
+                $originProduct->store_id = $product->store_from_id;
                 $originProduct->status = 'available';
                 $originProduct->save();
             }
